requirements.txt
# ----------------------------
# BACKEND & API
# ----------------------------
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
python-multipart>=0.0.6

# ----------------------------
# AI & AGENTS
# ----------------------------
langgraph>=0.0.10
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-ollama>=0.0.1

# ----------------------------
# VECTOR DATABASES
# ----------------------------
# Local Vector DB (optional)
# chromadb>=0.4.0
# langchain-chroma>=0.1.0

# Cloud Vector DB (Pinecone - STABLE)
pinecone>=6.0.0,<7.0.0
langchain-pinecone>=0.2.0

# ----------------------------
# DATA PROCESSING
# ----------------------------
pymupdf4llm>=0.0.1
langchain-text-splitters>=0.0.1

# ----------------------------
# MEMORY / DATABASE
# ----------------------------
sqlalchemy>=2.0.0
langgraph-checkpoint-sqlite

# ----------------------------
# UTILITIES
# ----------------------------
pyngrok>=7.0.0
nest_asyncio>=1.5.0

google-api-python-client google-auth google-auth-httplib2 google-auth-oauthlib

